+++ Practices/05.MVC/PractiveMVC/PracticeMVC/Controllers/HomeController.cs
@@ -1,13 +1,13 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using PractiveMVC.Models;
+using PracticeMVC.Models;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace PractiveMVC.Controllers
+namespace PracticeMVC.Controllers
 {
    public class HomeController : Controller
    {
