+++ Practices/05.MVC/PractiveMVC/PracticeMVC/Models/ErrorViewModel.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace PractiveMVC.Models
+namespace PracticeMVC.Models
 {
    public class ErrorViewModel
    {
