+++ Practices/05.MVC/PractiveMVC/PracticeMVC/Program.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace PractiveMVC
+namespace PracticeMVC
 {
    public class Program
    {
