+++ Practices/05.MVC/PractiveMVC/PracticeMVC/Views/Person/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model PractiveMVC.Models.PersonModel
+﻿@model PracticeMVC.Models.PersonModel
 
 @{
     ViewData["Title"] = "Edit";
