+++ Practices/05.MVC/PractiveMVC/PracticeMVC/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-﻿@using PractiveMVC
-@using PractiveMVC.Models
+﻿@using PracticeMVC
+@using PracticeMVC.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
